Always preserve first metadata value when reducing scores
Fixes #587

See conversation on PR #591

The simplest approach which is deterministic is to simply preserve the first metadata value when reducing (without any comparison). This will avoid the complicated comparison of metadata objects which have all the various traps we're running into and keep metadata around for common cases.

The alternative would be to never preserve score metadata, but this would have the effect of clearing the metadata even when a single score is being ‘reduced’ since we run the reducer pass no matter the number of epochs being run.